requirements.txt
aiosmtplib==4.0.2
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.31.0
bcrypt==3.2.2
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.122.0
fastapi-mail==1.5.8
h11==0.16.0
idna==3.11
iso8601==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.3
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pypika-tortoise==0.6.3
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
regex==2025.11.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tortoise-orm==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
